requirements.txt
pypdf==4.0.1
requests==2.31.0
anthropic==0.25.0
pandas==2.1.4
matplotlib==3.8.2
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
